+++ packages/message-kit/src/tests/Commands.test.ts
@@ -1,11 +1,11 @@
 import "dotenv/config";
-import { extractCommandValues } from "../helpers/utils";
+import { parseSkill } from "../lib/skills";
 import { skills } from "./Test_skills";
 
 describe("Command extraction tests", () => {
   test("Extract values from /tip command", () => {
     const inputContent = "/tip @bo @alix 15";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "tip",
       params: {
@@ -18,7 +18,7 @@ describe("Command extraction tests", () => {
   // /send 1 to @bo
   test("Extract values from /send command", () => {
     const inputContent = "/send 1 to @bo";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "send",
       params: {
@@ -31,7 +31,7 @@ describe("Command extraction tests", () => {
 
   test("Extract values from /swap command", () => {
     const inputContent = "/swap 10 eth to usdc";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "swap",
       params: {
@@ -44,7 +44,7 @@ describe("Command extraction tests", () => {
 
   test("Extract values from /send command", () => {
     const inputContent = "/send 10 usdc to @bo";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "send",
       params: {
@@ -57,7 +57,7 @@ describe("Command extraction tests", () => {
 
   test("Extract values from /send command", () => {
     const inputContent = "/send 10 usdc to @fabri";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "send",
       params: {
@@ -70,7 +70,7 @@ describe("Command extraction tests", () => {
 
   test("Extract values from /send command", () => {
     const inputContent = "/send 10 usdc vitalik.eth";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "send",
       params: {
@@ -83,7 +83,7 @@ describe("Command extraction tests", () => {
 
   test("Extract values from /game command", () => {
     const inputContent = "/game slot";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "game",
       params: {
@@ -94,7 +94,7 @@ describe("Command extraction tests", () => {
 
   test("Extract values from 🔎 emoji", () => {
     const inputContent = "🔎 ";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: undefined,
       params: {},
@@ -103,7 +103,7 @@ describe("Command extraction tests", () => {
 
   test("Extract values from /game help command", () => {
     const inputContent = "/game help";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "game",
       params: {
@@ -114,7 +114,7 @@ describe("Command extraction tests", () => {
 
   test("Extract values from /agent prompt command", () => {
     const inputContent = "/agent Hello, how can I assist you today?";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "agent",
       params: expect.objectContaining({
@@ -125,7 +125,7 @@ describe("Command extraction tests", () => {
 
   test("Extract values from /agent tell a joke command", () => {
     const inputContent = "/agent tell a joke";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "agent",
       params: {
@@ -136,7 +136,7 @@ describe("Command extraction tests", () => {
 
   test("Extract values from /send 1 eth to @bo command", () => {
     const inputContent = "/send 1 eth to @bo";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "send",
       params: {
@@ -149,7 +149,7 @@ describe("Command extraction tests", () => {
 
   test("Extract values from /show swap demo command", () => {
     const inputContent = "/show";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "show",
       params: {},
@@ -158,7 +158,7 @@ describe("Command extraction tests", () => {
 
   test("Extract values from /agent lets play a game command", () => {
     const inputContent = "/agent lets play a game";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "agent",
       params: {
@@ -169,7 +169,7 @@ describe("Command extraction tests", () => {
 
   test("Extract values from /help command", () => {
     const inputContent = "/help";
-    const extractedValues = extractCommandValues(inputContent, skills);
+    const extractedValues = parseSkill(inputContent, skills);
     expect(extractedValues).toEqual({
       command: "help",
       params: {},
